Allow to search on multilanguage sites.
lunrjs was trying to get the file index.json from the root while when the site is multilanguage there are more than one index.json, one per language and living within a folder called as the language.
This change tries to locate the file index.json according to the language specified in the URL of the current page or failing over the root path.